requirements.txt
Flask
tensorflow
Werkzeug
Pillow
gunicorn